+++ src/explorer/newFileWizard/actions.ts
@@ -20,10 +20,14 @@ export const newFileWizardShow = createAction(() => ({
  * Indicates that the use accepted the new file wizard dialog.
  * @param fileName The user-provided file name.
  * @param fileExtension The user-provided file extension.
- * @param hubType The user-provided hub type.
+ * @param hubType The user-provided hub type or undefined for an empty file.
  */
 export const newFileWizardDidAccept = createAction(
-    (fileName: string, fileExtension: SupportedFileExtension, hubType: Hub) => ({
+    (
+        fileName: string,
+        fileExtension: SupportedFileExtension,
+        hubType: Hub | undefined,
+    ) => ({
         type: 'explorer.newFileWizard.action.didAccept',
         fileName,
         fileExtension,